Write some tests and add refactor some code

+++ src/components/Tasks/Tasks.tsx
@@ -11,17 +11,25 @@ enum Mode {
 
 interface Props {
   tasks: Task[];
-  onChange: (task: Task) => void;
-  onDelete: (task: Task) => void;
+  onChange?: (task: Task) => void;
+  onDelete?: (task: Task) => void;
 }
 
 interface State {
   mode: Mode;
-  newTask?: Task;
+  activeItem: number;
+  newTask: Task;
 }
 
 class Tasks extends React.Component<Props, State> {
-  public state = { mode: Mode.default, newTask: createTask() };
+  constructor(props: Props) {
+    super(props);
+    this.state = {
+      mode: Mode.default,
+      newTask: createTask(),
+      activeItem: props.tasks.length - 1,
+    };
+  }
 
   public handleCreateTask = () => {
     this.setState({ mode: Mode.create, newTask: createTask() });
@@ -32,14 +40,32 @@ class Tasks extends React.Component<Props, State> {
   };
 
   public handleDeleteTask = (task: Task) => {
-    this.props.onDelete(task);
+    if (this.props.onDelete) {
+      this.setState({ activeItem: NaN }, () => {
+        if (this.props.onDelete) {
+          this.props.onDelete(task);
+        }
+      });
+    }
+  };
+
+  public handleExpand = (index: number, expand: boolean) => {
+    if (expand === false) {
+      return this.setState({ activeItem: NaN });
+    }
+    this.setState({ activeItem: index });
   };
 
   public createTask = (task: Task) => {
     if (isValidTask(task)) {
-      this.setState({ mode: Mode.default }, () => {
-        this.props.onChange(task);
-      });
+      this.setState(
+        { mode: Mode.default, activeItem: this.props.tasks.length },
+        () => {
+          if (this.props.onChange) {
+            this.props.onChange(task);
+          }
+        },
+      );
     }
   };
 
@@ -48,16 +74,23 @@ class Tasks extends React.Component<Props, State> {
     return (
       <>
         <ul className={styles.tasks}>
-          {tasks.map((t: Task) => (
+          {tasks.map((t: Task, index) => (
             <TaskItem
               key={t.id}
               task={t}
               onChange={onChange}
               onDelete={this.handleDeleteTask}
+              onExpand={(expand) => this.handleExpand(index, expand)}
+              confirmDelete={true}
+              expand={index === this.state.activeItem}
             />
           ))}
           {this.state.mode === Mode.create && (
-            <TaskItem task={this.state.newTask} onChange={this.createTask} />
+            <TaskItem
+              task={this.state.newTask}
+              onChange={this.createTask}
+              expand={true}
+            />
           )}
         </ul>
         {this.state.mode === Mode.default ? (
